+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # CHANGE LOG
 
-## 1.0.0
+## [1.0.1] - 2024-05-06
+
+### Added
+
+- Possibility for finding topics with Type & State Instance conditions.
+- Value Mapping to the MQTT Command Topic Interface.
+
+### Changed
+
+- Methods convertAliceValueToMqttMessage and convertMqttMessageToAliceValue are async now.
+
+### Removed
+
+- MqttConvertData type, and used CommandTopicData instead.
+
+## [1.0.0] - 2024-05-02
+
+### Added
 
 - Initial stable release.